requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
drf-spectacular==0.26.5
future==0.18.3
geojson==2.5.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
probableparsing==0.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pyowm==3.3.0
PySocks==1.7.1
python-crfsuite==0.9.9
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
